fix systemd service for escaped systemd paths
