+++ launcher/onedriver.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#define ONEDRIVER_SERVICE_TEMPLATE "onedriver@.service"
+
+/*
+ * Block until the filesystem becomes available.
+ */
+void poll_fs_availability(const char *mountpoint);
